move : Event 테스트 디렉토리 구조 및 클래스 이름 변경
